GPP - wip(MovSys): Rotation Snap
Snaps is integrated in rotation support values, and the visual representation is done, but it has currently no effect.
Note: It may be needed to call the UpdateSnapVisual from the construction function in blueprint, cause it seems that this function is called in situations for which I didn't found the c++ equivalent, and it should prevent snap arrows to disappear, especially when launching unreal.