GPP - add(MovSys): Interactable Handle
Added the MovSysHandle and its ability to control translation and rotation support.
Note: the MovSysHandle management of input does not take the camera view in its computations, so it is terrible with the free camera of the testing player.
It should be fine in the game since we have fixed camera angles