GPP - wip: Details panel customization discovery
This commit should never be merged inside the project, it's here only for educational purposes
Tests with the details panel customization, originally for setting up a multibox bool layout, but this has been disabled in unreal 5 haha. Still useful for the comprehension of how modules and details panel customization works.